@@ -1,5 +1,5 @@
 import React from 'react';
-import { Switch, Route, BrowserRouter} from 'react-router-dom';
+import { Routes, Route, BrowserRouter as Router} from 'react-router-dom';
 import { StylesProvider } from '@material-ui/core/styles';
 
 import Landing from './components/Landing'
@@ -8,12 +8,12 @@ import Pricing from './components/Pricing'
 export default () => {
   return <div>
       <StylesProvider>
-        <BrowserRouter>
-          <Switch>
-            <Route exact path="/pricing" component={Pricing} />
-            <Route path="/" component={Landing} />
-          </Switch>
-        </BrowserRouter>
+        <Router>
+          <Routes>
+            <Route path="/" element={<Landing />} />
+            <Route exact path="/pricing" element={<Pricing />} />
+          </Routes>
+        </Router>
       </StylesProvider>   
   </div>
 }
